==> Notebooks/0-Fetching Data.ipynb <==
import requests
import json
import pandas as pd 

# checking what we got
original = pd.read_csv('dialect_dataset.csv')
original.head()

# convert id to string because json needs strings not integers 
original['id_s'] = original['id'].astype(str)
original.head()

# first batch
t1 = original['id_s'][:197].tolist()
t1

# convert the batch to json to be used in request method
json_string = json.dumps(t1)
json_string

# getting respone in the shape of a dictionary
dict1 = requests.post("https://recruitment.aimtechnologies.co/ai-tasks", data= json_string).json()

# convert the dictionary to pandas dataframe 
dataframe = pd.DataFrame.from_dict(dict1, 'index')
dataframe

len(original.index)

# sanity check to make sure that we are not missing any index
t2 = original['id_s'].tolist()
len(t2)

# Sending request in chuncks starting from the first batch, then concatenating the dataframe to that of the first batch 
for i in range(0, len(t2), 1000):   
    json_string = json.dumps(t2[i+197:i+1197])
    dict1 = requests.post("https://recruitment.aimtechnologies.co/ai-tasks", data= json_string).json()
    dataframe1 = pd.DataFrame.from_dict(dict1, 'index')
    dataframe = pd.concat([dataframe, dataframe1]) 

# Use this CSV for preprocessing 
dataframe.to_csv('Data.csv')
