==> Notebooks/1-Merging Data.ipynb <==
import pandas as pd 

# original Dataset without text
original = pd.read_csv('dialect_dataset.csv', index_col=0)
original.size

# Dataset after fetching Data
api_re = pd.read_csv('Data.csv', index_col=0)
api_re.size

api_re

# merging in ID to make sure nothing is missing
to_work = original.join(api_re)
to_work

to_work.to_csv('Work.csv')
