==> Notebooks/2-Preprocessing Data.ipynb <==
import pandas as pd 
import re
import matplotlib.pyplot as plt

df = pd.read_csv('Work.csv', index_col='id')

# making sure there are no duplicates or nun values
df.describe()

# get users tags, links and hashtags
pattern1 = '@\S*|https\S*|#\S*'
# get more than 2 white spaces
pattern2 = '\s{2,}'
# get leading or ending white spaces
pattern3 = '^\s|\s$'

# before first pattern theres a link and a hashtag
trial_string = df['0'].values[246606]
trial_string

# first pattern is applied we get leading and ending whitespaces 
trial_string2 = re.sub(pattern1, '', trial_string)
trial_string2

# second pattern is applied we but there's no repeated white spaces so we dont observe the change :D, but we do observe the leading and ending white spaces
trial_string3 = re.sub('\s{2,}', ' ', trial_string2)
trial_string3

# third pattern is applied we get what we want
re.sub('^\s|\s$', '', trial_string3)

# apply these changes to the dataset
df['0'] = df['0'].str.replace(pattern1, '')
df['0'] = df['0'].str.replace(pattern2, ' ')
df['0'] = df['0'].str.replace(pattern3, '')

df.to_csv('cleaned.csv')
