==> Notebooks/5-Deep Learning local run.ipynb <==
from transformers import BertForSequenceClassification, TextClassificationPipeline, AutoTokenizer

tokenizer = AutoTokenizer.from_pretrained('aubmindlab/bert-base-arabertv02-twitter')

model = BertForSequenceClassification.from_pretrained("D:\\1-ITI-AI PRO\\DRAFT\\AIM ML\\trial10")

pipe = TextClassificationPipeline(model=model, tokenizer=tokenizer)

label_dict = {'LABEL_0' : 'AE', 'LABEL_1' : 'BH', 'LABEL_2' : 'DZ', 'LABEL_3' : 'EG', 'LABEL_4' : 'IQ', 'LABEL_5' : 'JO', 'LABEL_6' : 'KW',
              'LABEL_7' : 'LB', 'LABEL_8' : 'LY', 'LABEL_9' : 'MA', 'LABEL_10' : 'OM', 'LABEL_11' : 'PL',
              'LABEL_12' :'QA', 'LABEL_13' : 'SA', 'LABEL_14' : 'SD', 'LABEL_15' : 'SY', 'LABEL_16' : 'TN', 'LABEL_17' : 'YE'}

s = 'ليه بس كده يا باشا' 

for pred in pipe(s):
    print(label_dict[pred['label']])
