==> Notebooks/7-Deep Learning Flask.ipynb <==
import numpy as np
from flask import Flask, request, jsonify, render_template
from transformers import BertForSequenceClassification, TextClassificationPipeline, AutoTokenizer

app = Flask(__name__)

tokenizer = AutoTokenizer.from_pretrained('aubmindlab/bert-base-arabertv02-twitter')
model = BertForSequenceClassification.from_pretrained("D:\\1-ITI-AI PRO\\DRAFT\\AIM ML\\trial10") # where the model is saved
pipe = TextClassificationPipeline(model=model, tokenizer=tokenizer)

label_dict = {'LABEL_0' : 'AE', 'LABEL_1' : 'BH', 'LABEL_2' : 'DZ', 'LABEL_3' : 'EG', 'LABEL_4' : 'IQ', 'LABEL_5' : 'JO',
              'LABEL_6' : 'KW', 'LABEL_7' : 'LB', 'LABEL_8' : 'LY', 'LABEL_9' : 'MA', 'LABEL_10' : 'OM', 'LABEL_11' : 'PL',
              'LABEL_12' :'QA', 'LABEL_13' :' SA', 'LABEL_14' : 'SD', 'LABEL_15' : 'SY', 'LABEL_16' : 'TN', 'LABEL_17' : 'YE'}

@app.route('/')
def home():
    return render_template('index.html')

@app.route('/predict',methods=['POST'])
def predict():
    '''
    For rendering results on HTML GUI
    '''
    int_features = [str(x) for x in request.form.values()]    
    prediction = pipe(int_features[0])[0]
    output = label_dict[prediction['label']]

    return render_template('index.html', prediction_text='Expected Dialect: {}'.format(output))

@app.route('/predict_api',methods=['POST'])
def predict_api():
    '''
    For direct API calls through request
    '''
    data = request.get_json(force=True)    
    prediction = pipe(list(data.values()))    
    output = [label_dict[pred['label']] for pred in prediction]
    
    return jsonify(output)


if __name__ == "__main__":
    app.run(debug=False)
